==> analysis/03-fig4.Rmd <==
---
title: "Figure 4"
---

```{r knitr, include = FALSE}
DOCNAME = "Figure-4"
NOW <- Sys.time()

# Time chunks during knitting
knitr::knit_hooks$set(timeit = function(before) {
    if (before) {
        print(paste("Start:", Sys.time()))
        NOW <<- Sys.time()
    } else {
        print(paste("Stop:", Sys.time()))
        print(Sys.time() - NOW)
    }
})

knitr::opts_chunk$set(
    autodep        = TRUE,
    cache          = TRUE,
    cache.path     = paste0("cache/", DOCNAME, "/"),
    cache.comments = FALSE,
    echo           = TRUE,
    error          = FALSE,
    fig.align      = "center",
    fig.width      = 10,
    fig.height     = 8,
    message        = FALSE,
    warning        = FALSE,
    timeit         = TRUE
)
```

```{r libaries, cache = FALSE}
# scRNA-seq
library(Seurat)
packageVersion("Seurat")

# Plotting
library("ggplot2")
library("ggpubr")
library("cowplot")

# Presentation
library("knitr")

# Others
library("readxl")
library("DT")
library("dplyr")
library('stringr')
```

```{r source, cache = FALSE}
source(here::here("R/00_generalDeps.R"))
source(here::here("R/output.R"))
```

```{r depends-paths}
filt_path <- here::here("data/processed/patient.integrated.rds")
```

Introduction
============

In this document we are going to perform plotting figure4 panelC&D on the high-quality filtered 
dataset using `Seurat`.


Loading `seurat` object processed
```{r load, cache.extra = tools::md5sum(filt_path)}
if (file.exists(filt_path)) {
    patient.integrated <- readRDS(filt_path)
} else {
    stop("Figure4 dataset is missing. ",
         "Please check Input first.",
         call. = FALSE)
}
```

Integration  
==============

Integrate data from the six samples, so that cells from the same cell type/subpopulation will cluster together.

visualization
------

plot metadata

```{r plot-data}
#DimPlot(patient.integrated, reduction = "umap", group.by = "SortPheno") + ggtitle(paste0("SortPheno: ", length(patient.integrated$SortPheno)))

clusterCount = as.data.frame( table( ClusterNumber = patient.integrated[[ "SortPheno" ]]), responseName = "CellCount");
# Create datatable
datatable( clusterCount, 
           class = "compact",
           rownames = FALSE,
           colnames = c("SortPheno", "Nb. Cells"),
           options = list(dom = "<'row'rt>", # Set elements for CSS formatting ('<Blf><rt><ip>')
                          autoWidth = FALSE,
                          columnDefs = list( # Center all columns
                            list( targets = 0:(ncol(clusterCount)-1),
                                  className = 'dt-center')),
                          orderClasses = FALSE, # Disable flag for CSS to highlight columns used for ordering (for performance)
                          paging = FALSE, # Disable pagination (show all)
                          processing = TRUE, 
                          scrollCollapse = TRUE,
                          scroller = TRUE,  # Only load visible data
                          scrollX = TRUE,
                          scrollY = "525px",
                          stateSave = TRUE));

#DimPlot(patient.integrated, reduction = "umap", group.by = "Sample")  + ggtitle(paste0("Sample: ", length(patient.integrated$Sample)))

clusterCount = as.data.frame( table( ClusterNumber = patient.integrated[[ "Sample" ]]), responseName = "CellCount");
# Create datatable
datatable( clusterCount, 
           class = "compact",
           rownames = FALSE,
           colnames = c("Sample", "Nb. Cells"),
           options = list(dom = "<'row'rt>", # Set elements for CSS formatting ('<Blf><rt><ip>')
                          autoWidth = FALSE,
                          columnDefs = list( # Center all columns
                            list( targets = 0:(ncol(clusterCount)-1),
                                  className = 'dt-center')),
                          orderClasses = FALSE, # Disable flag for CSS to highlight columns used for ordering (for performance)
                          paging = FALSE, # Disable pagination (show all)
                          processing = TRUE, 
                          scrollCollapse = TRUE,
                          scroller = TRUE,  # Only load visible data
                          scrollX = TRUE,
                          scrollY = "525px",
                          stateSave = TRUE));

```

Plate4 01-018 issue
------

Visual plate4 sample 01-018 then filter out

```{r filtering-out}

DefaultAssay(patient.integrated) <- 'RNA'
metadata_df = patient.integrated@meta.data
ids = metadata_df[metadata_df$Plate == 'p4' & metadata_df$Sample == '01-018',]$UniqueCellID

metadata_df$p4018 = 'no'   
metadata_df[ids, ]$p4018 = 'yes'
patient.integrated$p4018 = metadata_df$p4018 
#DimPlot(patient.integrated, reduction = "umap", group.by = "p4018")  + ggtitle("Plate 4 018 issue") 

Idents(patient.integrated) <- 'p4018'
patient.integrated = subset(patient.integrated, idents = 'yes',invert = TRUE)

```
visualization 2
------

plot metadata after filter out
```{r plot-data2}

#DimPlot(patient.integrated, reduction = "umap", group.by = "SortPheno") + ggtitle(paste0("After filtering Plate 4 018 issue SortPheno: ", length(patient.integrated$SortPheno)))
clusterCount = as.data.frame( table( ClusterNumber = patient.integrated[[ "SortPheno" ]]), responseName = "CellCount");
# Create datatable
datatable( clusterCount, 
           class = "compact",
           rownames = FALSE,
           colnames = c("SortPheno", "Nb. Cells"),
           options = list(dom = "<'row'rt>", # Set elements for CSS formatting ('<Blf><rt><ip>')
                          autoWidth = FALSE,
                          columnDefs = list( # Center all columns
                            list( targets = 0:(ncol(clusterCount)-1),
                                  className = 'dt-center')),
                          orderClasses = FALSE, # Disable flag for CSS to highlight columns used for ordering (for performance)
                          paging = FALSE, # Disable pagination (show all)
                          processing = TRUE, 
                          scrollCollapse = TRUE,
                          scroller = TRUE,  # Only load visible data
                          scrollX = TRUE,
                          scrollY = "525px",
                          stateSave = TRUE));

#DimPlot(patient.integrated, reduction = "umap", group.by = "Sample")  + ggtitle(paste0("After filtering Plate 4 018 issue Sample: ", length(patient.integrated$Sample)))

clusterCount = as.data.frame( table( ClusterNumber = patient.integrated[[ "Sample" ]]), responseName = "CellCount");
# Create datatable
datatable( clusterCount, 
           class = "compact",
           rownames = FALSE,
           colnames = c("Sample", "Nb. Cells"),
           options = list(dom = "<'row'rt>", # Set elements for CSS formatting ('<Blf><rt><ip>')
                          autoWidth = FALSE,
                          columnDefs = list( # Center all columns
                            list( targets = 0:(ncol(clusterCount)-1),
                                  className = 'dt-center')),
                          orderClasses = FALSE, # Disable flag for CSS to highlight columns used for ordering (for performance)
                          paging = FALSE, # Disable pagination (show all)
                          processing = TRUE, 
                          scrollCollapse = TRUE,
                          scroller = TRUE,  # Only load visible data
                          scrollX = TRUE,
                          scrollY = "525px",
                          stateSave = TRUE));
```



Box plots of gene set score
==============

loading markers table
------

markers table 1 : figure 2, marker table 2 : figure 3

```{r loading-table}
# DEA by Anno_1
dea_f = read.csv(file = here::here("data/processed/DEA_byAnno_1.csv"), header = TRUE)
Top50_dea_f = dea_f %>% group_by(cluster) %>% top_n(n = 50, wt = avg_log2FC)
selfAg_df <- Top50_dea_f[Top50_dea_f$cluster == "SLA",]
flu_df <- Top50_dea_f[Top50_dea_f$cluster == "FLU",]

#marker2
amr053_df = read_excel(here::here("data/processed/supp_Table7_Rev1.xlsx"), sheet = "Feuil1")
amr053_df = na.omit(amr053_df)
```

Box plots
------
Gene set score Auto-reactivity & H1N1

```{r Boxplots}
for(name in c('selfAg', 'flu')){
  df = eval(parse(text = paste0(name, '_df')))
  patient.integrated <- AddModuleScore(patient.integrated, features = list(df$gene) ,name = name)
}

metadata_df = patient.integrated@meta.data

title = c("Auto-reactivity", "H1N1")
i = 1
for(name in c('selfAg1','flu1')){
    p<-ggplot(metadata_df, aes(x=SortPheno, y= eval(parse(text = name)), color=Sample)) +
      geom_boxplot()+ 
      #scale_color_manual(values=cols) + 
      geom_jitter(color="black", size=0.4, alpha=0.9) + 
      theme_bw() +
      theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1))
    
    my_comparisons <- list(c("TTneg" ,  "TTpos"))
    
    p <- p + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
    
    if ( i == 1 ){
        
      stat = compare_means(selfAg1 ~ SortPheno, data = metadata_df)

    }else{
      
      stat = compare_means(flu1 ~ SortPheno, data = metadata_df)

    }
    
    print(p + ggtitle(paste0(title[i],": p.format is ", stat$p.format)))
    i = i + 1
}
```

Cluster markers
-----------------------

```{r cluster-marker}
for(name in c(1:3,5,6)){
  genes = amr053_df[amr053_df$cluster == name,]$gene
  
  if(length(genes) > 50){
    genes = genes[1:50]
  }
  patient.integrated <- AddModuleScore(patient.integrated, features = list(genes) ,name = paste0('cl', name))
}

metadata_df = patient.integrated@meta.data


title = c("Cluster 1 (TEM GZMK+)", "Cluster 2 (TCM HLA-DR+)", "Cluster 3 (TEM NKG7+)", "Cluster 5 (CTL)", "Cluster 6 (TREG)")
i = 1
for(name in paste0('cl', c(1:3,5,6), '1')){
  p<-ggplot(metadata_df, aes(x=SortPheno, y= eval(parse(text = name)), color=SortPheno)) +
    geom_boxplot()+ 
    #scale_color_manual(values=cols) + 
    geom_jitter(color="black", size=0.4, alpha=0.9) + 
    theme_bw() +
    theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1)) 
  
  my_comparisons <- list(c("TTneg" ,  "TTpos"))
  
  p <- p + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
  
   if ( i == 1 ){
        
      stat = compare_means(cl11 ~ SortPheno, data = metadata_df)

    }else if (i == 2){
      
      stat = compare_means(cl21 ~ SortPheno, data = metadata_df)

    }else if (i == 3){
      
      stat = compare_means(cl31 ~ SortPheno, data = metadata_df)

    }else if (i == 4){
      
      stat = compare_means(cl51 ~ SortPheno, data = metadata_df)

    }else if (i == 5){
      
      stat = compare_means(cl61 ~ SortPheno, data = metadata_df)

    }
    print(p + ggtitle(paste0(title[i],": p.format is ", stat$p.format)))
    i = i + 1
}
```

Figure 4
=======

Panel C
--------------

```{r fig4-pC}

my_comparisons <- list(c("TTneg" ,  "TTpos"))
p1 <-ggplot(metadata_df, aes(x=SortPheno, y= selfAg1, color=Sample)) +
      geom_boxplot()+ 
      #scale_color_manual(values=cols) + 
      geom_jitter(color="black", size=0.4, alpha=0.9) + 
      theme_bw() +
      theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1))
    
p1 <- p1 + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
p1 <- p1 + labs(x = "sorting phenotype", y = "Gene set score", title = "Auto-reactivity")

p2 <-ggplot(metadata_df, aes(x=SortPheno, y= flu1, color=Sample)) +
      geom_boxplot()+ 
      #scale_color_manual(values=cols) + 
      geom_jitter(color="black", size=0.4, alpha=0.9) + 
      theme_bw() +
      theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1))
    
p2 <- p2 + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
p2 <- p2 + labs(x = "sorting phenotype", y = "Gene set score", title = "H1N1")

fig <- plot_grid(p1, p2, nrow = 1, labels = "AUTO")

ggsave(here::here("output", DOCNAME, "figure4-panelC.pdf"), fig,
        width = 14, height = 8, scale = 1.2)
ggsave(here::here("output", DOCNAME, "figure4-panelC.png"), fig,
        width = 14, height = 8, scale = 1.2)

fig
```

Panel D
----------
```{r fig4-pD}

p1<-ggplot(metadata_df, aes(x=SortPheno, y= cl11, color=SortPheno)) +
    geom_boxplot()+ 
    geom_jitter(color="black", size=0.4, alpha=0.9) + 
    theme_bw() +
    theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1)) 
p1 <- p1 + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
p1 <- p1 + labs(x = "sorting phenotype", y = "Gene set score", title = "Cluster 1 (TEM GZMK+)")

p2<-ggplot(metadata_df, aes(x=SortPheno, y= cl21, color=SortPheno)) +
    geom_boxplot()+ 
    geom_jitter(color="black", size=0.4, alpha=0.9) + 
    theme_bw() +
    theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1)) 
p2 <- p2 + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
p2 <- p2 + labs(x = "sorting phenotype", y = "Gene set score", title = "Cluster 2 (TCM HLA-DR+)")

p3<-ggplot(metadata_df, aes(x=SortPheno, y= cl31, color=SortPheno)) +
    geom_boxplot()+ 
    geom_jitter(color="black", size=0.4, alpha=0.9) + 
    theme_bw() +
    theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1)) 
p3 <- p3 + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
p3 <- p3 + labs(x = "sorting phenotype", y = "Gene set score", title = "Cluster 3 (TEM NKG7+)")

p4<-ggplot(metadata_df, aes(x=SortPheno, y= cl51, color=SortPheno)) +
    geom_boxplot()+ 
    geom_jitter(color="black", size=0.4, alpha=0.9) + 
    theme_bw() +
    theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1)) 
p4 <- p4 + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
p4 <- p4 + labs(x = "sorting phenotype", y = "Gene set score", title = "Cluster 5 (CTL)")

p5<-ggplot(metadata_df, aes(x=SortPheno, y= cl61, color=SortPheno)) +
    geom_boxplot()+ 
    geom_jitter(color="black", size=0.4, alpha=0.9) + 
    theme_bw() +
    theme(axis.text.x = element_text(angle = 0, vjust = 1, hjust=1)) 
p5 <- p5 + stat_compare_means(comparisons = my_comparisons, label = "p.signif")
p5 <- p5 + labs(x = "sorting phenotype", y = "Gene set score", title = "Cluster 6 (Treg)")


fig <- plot_grid(p1, p2, p3, p4, p5, nrow = 2, labels = "AUTO")

ggsave(here::here("output", DOCNAME, "figure4-panelD.pdf"), fig,
       width = 14, height = 8, scale = 1.3)
ggsave(here::here("output", DOCNAME, "figure4-panelD.png"), fig,
       width = 14, height = 8, scale = 1.3)

fig
```

Output files
==============

This table describes the output files produced by this document. Right click
and _Save Link As..._ to download the results.

```{r save}
saveRDS(patient.integrated, here::here("data/processed/03-sc_seurat.rds"))
save(metadata_df, file = here::here("data/processed/03-sc_seurat_metadata.RData"))
```


```{r output}
dir.create(here::here("output", DOCNAME), showWarnings = FALSE)

knitr::kable(data.frame(
     File = c(
         getDownloadLink2("figure4-panelC.png", str_sub(here::here("output", DOCNAME),2,-1)), # bug one "/" plus
         getDownloadLink2("figure4-panelC.pdf", str_sub(here::here("output", DOCNAME),2,-1)),
         getDownloadLink2("figure4-panelD.png", str_sub(here::here("output", DOCNAME),2,-1)),
         getDownloadLink2("figure4-panelD.pdf", str_sub(here::here("output", DOCNAME),2,-1))
     ),
     Description = c(
         "Figure 4 panel C (PNG)",
         "Figure 4 panel C (PDF)",
         "Figure 4 panel D (PNG)",
         "Figure 4 panel D (PDF)"
     )
 ))
```


Session information
==============

```{r session-info, cache = FALSE}
devtools::session_info()
```
