==> analysis/06-suppfig4.Rmd <==
---
title: "Supplementary figure 4"
---

```{r knitr, include = FALSE}
DOCNAME = "Suppfigure-4"
NOW <- Sys.time()

# Time chunks during knitting
knitr::knit_hooks$set(timeit = function(before) {
    if (before) {
        print(paste("Start:", Sys.time()))
        NOW <<- Sys.time()
    } else {
        print(paste("Stop:", Sys.time()))
        print(Sys.time() - NOW)
    }
})

knitr::opts_chunk$set(
    autodep        = TRUE,
    cache          = TRUE,
    cache.path     = paste0("cache/", DOCNAME, "/"),
    cache.comments = FALSE,
    echo           = TRUE,
    error          = FALSE,
    fig.align      = "center",
    fig.width      = 10,
    fig.height     = 8,
    message        = FALSE,
    warning        = FALSE,
    timeit         = TRUE
)
```

```{r libaries, cache = FALSE}
# scRNA-seq
library(Seurat)
#packageVersion("Seurat")

# Plotting
library("ggplot2")
library(cowplot)

# Presentation
library("knitr")

# Others
library(stringr)
library("readxl")
library("rstatix")
library(ggpubr)
library('scales')
library( "RColorBrewer")

```

```{r source, cache = FALSE}
source(here::here("R/00_generalDeps.R"))
source(here::here("R/output.R"))
```

```{r depends-paths}
filt_path <- here::here("data/processed/02-sc_seurat_subset.rds")
```

Introduction
============

In this document we are going to perform plotting supplementary figure 4 panel A 
using `Seurat`.


Loading `seurat` object processed
```{r load, cache.extra = tools::md5sum(filt_path)}
if (file.exists(filt_path)) {
    subset <- readRDS(filt_path)
} else {
    stop("Supplementary figure 4  dataset is missing. ",
         "Please check Input first.",
         call. = FALSE)
}
```


```{r, echo=FALSE,include = FALSE}
# You need this code to conduct the magic dependences attaching...
DT::datatable(matrix())
```

visualization  
==============

Literature-based-gene set score

```{r plot-data, echo=TRUE, message=FALSE, warning=FALSE, results='asis'}

DimPlot( subset, label=TRUE, label.size = 8) + 
         theme( axis.title.x = element_blank(),
                  axis.title.y = element_blank(),
                  legend.position = "none") 

gene_set_pub <- read_excel(paste0(here::here("data/processed/Gene set publications.xlsx")), col_names = FALSE, sheet = "Feuil1")
gene_set_pub <- as.data.frame(gene_set_pub)


# stocke in list
boxplot_list = list()
featureplot_list = list()
df_scores <- subset@meta.data[,c("UniqueCellID","Group")]

for( j in c(3:5,10,12,15,16)){
  paper_names = paste0(gene_set_pub[1,j],"_",gene_set_pub[3,j], "_",gene_set_pub[4,j], "_",gene_set_pub[5,j])
  paper_names = gsub('\\(','',paper_names)
  paper_names = gsub('\\)','',paper_names)
  paper_names = gsub(' ','_',paper_names)
  cat("###",paper_names, '<br>', '\n')

  genes =   gene_set_pub[11:nrow(gene_set_pub),j]
  # Remove NA
  genes <- genes[!is.na(genes)]
  # filtering
  genes = genes[genes %in% rownames(subset)]

  subset <- AddModuleScore(subset,
                              features = list(genes),
                              ctrl = 5,
                              name = paper_names)

  df <- subset@meta.data
  
  # List module score  violin Plot pairwise wilcoxon test
  liste_scores <- df[,c(paste0(paper_names, "1"),'Group')]
  names(liste_scores) <- c('mod_score','Group')
  pwWlcx_scores <- pairwise_wilcox_test(liste_scores,formula=mod_score ~ Group,p.adjust.method = "BH")
  
  # Create datatable
  cat(knitr::knit_print(DT::datatable( as.data.frame(pwWlcx_scores), height = "100%", width = "100%",
                                       class = "compact",
                                       rownames = FALSE,
                                       #colnames = c("Sample", "Nb. Cells"),
                                       options = list(dom = "<'row'rt>", # Set elements for CSS formatting ('<Blf><rt><ip>')
                                                      autoWidth = FALSE,
                                                      columnDefs = list( # Center all columns
                                                        list( targets = 0:(ncol(pwWlcx_scores)-1),
                                                              className = 'dt-center')),
                                                      orderClasses = FALSE, # Disable flag for CSS to highlight columns used for ordering (for performance)
                                                      paging = FALSE, # Disable pagination (show all)
                                                      processing = TRUE,
                                                      scrollCollapse = TRUE,
                                                      scroller = TRUE,  # Only load visible data
                                                      scrollX = TRUE,
                                                      scrollY = "525px",
                                                      stateSave = TRUE))
  ))
  
  p <- ggplot(df, aes(x=Group, y=eval(parse(text = paste0(paper_names, "1"))), color =Group)) +
    geom_boxplot(color="grey") + geom_jitter(shape=16, position=position_jitter(0.2))+
    theme_classic() +
    stat_pvalue_manual(pwWlcx_scores, hide.ns = FALSE, y.position=max(df[[paste0(paper_names, "1")]])+0.05, step.increase=0.05) +
    #scale_color_manual(values=col)+
    theme(axis.text.x = element_text(angle = 0)) + theme(legend.position = 'none') + xlab("") + ylab(paper_names)
  print(p)
  
  # variable scope: local
  boxplot_list[[paper_names]] <- local({
        paper_names <- paper_names
        ggplot(df, aes(x=Group, y=eval(parse(text = paste0(paper_names, "1"))), color =Group)) +
    geom_boxplot(color="grey") + geom_jitter(shape=16, position=position_jitter(0.2))+
    theme_classic() +
    stat_pvalue_manual(pwWlcx_scores, hide.ns = FALSE, y.position=max(df[[paste0(paper_names, "1")]])+0.05, step.increase=0.05) +
    #scale_color_manual(values=col)+
    theme(axis.text.x = element_text(angle = 0)) + theme(legend.position = 'none') + xlab("") + ylab(paper_names)
    })
  
  
  #print(boxplot_list[[paper_names]] )
  
  p<-FeaturePlot(subset,
                 features = paste0(paper_names, "1"), repel = TRUE) +
    scale_colour_gradientn(colours = rev(brewer.pal(n = 11, name = "RdBu")))

  print(p)
  
  featureplot_list[[paper_names]] <- p +  labs(title = "")& NoAxes()
  
  df_scores[[paper_names]] = subset@meta.data[[paste0(paper_names, "1")]]
  
  cat("\n\n")
}

```


Supplementary figure 4
=======

Panel A
--------------

```{r suppfig4-pA, fig.width = 14, fig.height = 10}

fig <- plot_grid(boxplot_list[[1]], featureplot_list[[1]],
                 boxplot_list[[2]], featureplot_list[[2]],
                 boxplot_list[[3]], featureplot_list[[3]],
                 boxplot_list[[4]], featureplot_list[[4]],
                 boxplot_list[[5]], featureplot_list[[5]],
                 boxplot_list[[6]], featureplot_list[[6]],
                 boxplot_list[[7]], featureplot_list[[7]],
                 nrow = 3, labels = "A")

dir.create(here::here("output", DOCNAME), showWarnings = FALSE)

ggsave(here::here("output", DOCNAME, "suppfigure4-panelA.pdf"), fig,
        width = 24, height = 12, scale = 1)
ggsave(here::here("output", DOCNAME, "suppfigure4-panelA.png"), fig,
        width = 24, height = 12, scale = 1)

fig

```


Output files
==============

This table describes the output files produced by this document. Right click
and _Save Link As..._ to download the results.

```{r save}
saveRDS(subset, here::here("data/processed/06-sc_seurat_subset.rds"))
write.table(df_scores, here::here("output", DOCNAME,"06-raw_data.csv"), row.names=TRUE,col.names=TRUE, sep =',')
```


```{r output}
dir.create(here::here("output", DOCNAME), showWarnings = FALSE)

knitr::kable(data.frame(
     File = c(
         getDownloadLink2("06-raw_data.csv", str_sub(here::here("output", DOCNAME),2,-1)),
         getDownloadLink2("suppfigure4-panelA.png", str_sub(here::here("output", DOCNAME),2,-1)), # bug one "/" plus
         getDownloadLink2("suppfigure4-panelA.pdf", str_sub(here::here("output", DOCNAME),2,-1))
     ),
     Description = c(
         "raw data for source_data (CSV)",
         "Supplementary Figure 4 panel A (PNG)",
         "Supplementary Figure 4 panel A (PDF)"
     )
 ))
```


Session information
==============

```{r session-info, cache = FALSE}
devtools::session_info()
```
